GL-385: refactor: move check_moving_requirements_form under green_lanes module
