+++ spec/forms/green_lanes/check_moving_requirements_form_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-RSpec.describe CheckMovingRequirementsForm, type: :model do
+RSpec.describe GreenLanes::CheckMovingRequirementsForm, type: :model do
   describe 'validations' do
     subject(:form) { described_class.new(params) }
